+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/SpanFactory.kt
@@ -18,6 +18,8 @@ import com.bugsnag.android.performance.internal.integration.NotifierIntegration
 import com.bugsnag.android.performance.internal.metrics.MetricsContainer
 import com.bugsnag.android.performance.internal.processing.AttributeLimits
 import com.bugsnag.android.performance.internal.processing.SpanTaskWorker
+import com.bugsnag.android.performance.internal.util.Prioritized
+import com.bugsnag.android.performance.internal.util.PrioritizedSet
 import java.util.UUID
 
 internal typealias AttributeSource = (target: SpanImpl) -> Unit
@@ -35,7 +37,7 @@ public class SpanFactory internal constructor(
 
     internal var attributeLimits: AttributeLimits? = null
 
-    internal var spanStartCallbacks: Array<OnSpanStartCallback> = emptyArray()
+    internal val spanStartCallbacks = PrioritizedSet<OnSpanStartCallback>()
 
     public constructor(
         spanProcessor: SpanProcessor,
@@ -54,14 +56,15 @@ public class SpanFactory internal constructor(
     internal fun configure(
         spanProcessor: SpanProcessor,
         attributeLimits: AttributeLimits,
-        spanStartCallbacks: Array<OnSpanStartCallback>,
+        spanStartCallbacks: Collection<Prioritized<OnSpanStartCallback>>,
         networkRequestCallback: NetworkRequestInstrumentationCallback?,
         enabledMetrics: EnabledMetrics,
     ) {
         this.spanProcessor = spanProcessor
         this.attributeLimits = attributeLimits
-        this.spanStartCallbacks = spanStartCallbacks
         this.networkRequestCallback = networkRequestCallback
+
+        this.spanStartCallbacks.addAll(spanStartCallbacks)
 
         metricsContainer.configure(enabledMetrics)
     }