+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.16.0 (2025-06-23)
+
+### Changes
+
+* Added `Plugin` as a way to extend the functionality of the BugsnagPerformance library
+  [#387](https://github.com/bugsnag/bugsnag-android-performance/pull/387)
+* Introduced `AppStartSpanControl` and `SpanType.AppStart` to allow custom categorisation of app starts
+  [#394](https://github.com/bugsnag/bugsnag-android-performance/pull/394)
+
 ## 1.15.0 (2025-06-17)
 
 ### Changes